69. Sqrt(x)
